Add a post process event handler to perform error/debug handling (#59)
* Add more error handling to flow states + debuging options

* Debug logs: Split logs by date folders

* Allow more post process events + some renaming

* Set failed alingment to error + save failed alignment results

* Align file retention functions

* Reset process error at process start

* OTA update: Fully allow firmware.bin

* Adapt camera init + continue at any cam error to regular interface

* Continue to regular interface at any cam error

* Adapt cam deinit + update demo mode

* Modify doInit to prepare for reloading cam driver

* Add SPIRAM coverage info block

* Reload cam driver after error detected during processing

+++ code/components/jomjol_flowcontroll/ClassFlowImage.h
@@ -12,7 +12,7 @@ class ClassFlowImage : public ClassFlow
 protected:
 	std::string imagesLocation;
     bool isLogImage;
-    unsigned short imagesRetention;
+    int imagesRetention;
 	const char* logTag;
 
 	std::string CreateLogFolder(std::string time);
